requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
django-braces==1.17.0
django-debug-toolbar==6.2.0
django-embed-video==1.4.10
django-redisboard==9.0.0
djangorestframework==3.16.1
idna==3.11
pillow==12.1.1
pymemcache==4.0.0
redis==7.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
